requirements.txt
Django==4.2.7
Pillow==10.1.0
python-dotenv==1.0.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-filter==23.3
django-widget-tweaks==1.5.0
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-storages==1.14.2
django-cors-headers==4.3.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
django-cleanup==8.0.0
django-allauth==0.57.0
django-redis==5.4.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.32.0
django-prometheus==2.3.1
django-health-check==3.17.0
django-ratelimit==4.1.0
django-simple-history==3.4.0
django-import-export==3.3.1
django-guardian==2.4.0
django-otp==1.2.2
django-two-factor-auth==1.15.5
django-axes==6.2.0
django-password-validators==1.3.0
django-constance==2.9.1
django-maintenance-mode==0.20.0
django-admin-honeypot==1.1.0
django-admin-interface==0.28.0
django-colorfield==0.10.0
django-ckeditor==6.7.0
django-tinymce==3.7.0
django-summernote==0.8.20.0
django-filebrowser==4.0.3
django-grappelli==3.0.5
django-jet-reboot==0.2.8
django-jazzmin==2.6.0
django-admin-rangefilter==0.12.0
django-admin-list-filter-dropdown==1.0.3
django-admin-autocomplete-filter==0.6.2
django-admin-sortable2==2.1.13
django-admin-actions==1.1.1
django-admin-tools==0.9.3
django-admin-interface==0.28.0
django-admin-honeypot==1.1.0
django-admin-rangefilter==0.12.0
django-admin-list-filter-dropdown==1.0.3
django-admin-autocomplete-filter==0.6.2
django-admin-sortable2==2.1.13
django-admin-actions==1.1.1
django-admin-tools==0.9.3 